Issue #6 links should open new tabs
A few updates:

Links
----
Modified the links throughout the help pages, such that when you
click them, it loads them into a new tab. I think is appropriate,
as users are wanting to be on the help pages, and not be continuously
redirected to external sites constantly having to press back.

I would recommend keeping to this so that they do not have to be
updated in the future.

For now, I have left the blogs untouched, as I do not see them directly
related to the help pages.
---

Other:

 - Updated links to references to use bumblebee and arxiv/ADS2.0
 - Removed any links pointing to old help pages, when possible
 - Minor typos/styling of references.