Issue #7 edit this page on github
Added a small (temporary) link in the navbar that says you can edit the pages
on the github repo.

Also added a contribuing file, that has some simple guidelines on making
changes to the helpe pages. This can be made more simple, but given no one will have access to make direct changes to the pages, they'll have to make a PR
regardless.