Update 2016-07-29-comprehensive-solr-term-list.md
corrected typos and improved wording of some descriptions.  I know this was a table that Jonny made but why are we including the field "useful to users" in the help pages?  If we are showing the options to users than the user should be able to decide whether or not it's useful to them.